requirements.txt
fastapi==0.115.14
uvicorn==0.15.0
openai==1.97.1
python-dotenv==1.1.1
psycopg2-binary==2.9.10
cloudinary==1.44.1
sqlalchemy==2.0.42
pydantic==2.11.7
h11==0.16.0
pip==24.3.1
wheel==0.37.1
protobuf==5.28.2
cryptography==43.0.3
requests==2.32.3
colorama==0.4.4
typing_extensions==4.14.1
exceptiongroup==1.3.0
sniffio==1.3.1
click==8.0.3
httpcore==1.0.9
idna==3.3
httpx==0.28.1
jiter==0.10.0
distro==1.7.0
anyio==4.9.0
pydantic_core==2.33.2
certifi==2020.6.20
starlette==0.46.2
urllib3==1.26.5
wsproto==1.0.0
greenlet==3.2.3
annotated-types==0.7.0
PyYAML==5.4.1
six==1.16.0
setuptools==44.1.1
importlib-metadata==4.6.4